requirements.txt
openai-whisper
audioop-lts
PyQt6
sounddevice
soundfile
pydub
deep-translator
numpy
pyinstaller
